#99 WIP adding integration tests
