+++ api-client/src/api/apiClient.ts
@@ -2,7 +2,7 @@ import axios from "axios";
 import * as dotenv from "dotenv"
 dotenv.config()
 
-const axiosClient = axios.create({
+export const axiosClient = axios.create({
     baseURL: process.env.API_BASE_PATH,
     headers: {
         'Accept': 'application/json',
