+++ api-client/src/api/requestHandlers/quizRequest.ts
@@ -1,0 +1,5 @@
+import {axiosClient} from "../apiClient";
+
+export function getAll() {
+    return axiosClient.get('/quiz')
+}
