+++ api-client/tests/unit/integration/api/requestHandlers/quizRequest.test.ts
@@ -1,0 +1,26 @@
+import {getAll} from "../../../../../src/api/requestHandlers/quizRequest";
+import {mockApp} from "../../../../../src/api/mockClient";
+
+// Start the JSONServer
+const app = mockApp()
+
+beforeAll(() => {
+
+});
+
+afterAll(() => {
+    // Stop the JSONServer
+    app.close()
+});
+
+describe('API Request /quiz', () => {
+    it('Should have a status code of 200', async () => {
+        const response = await getAll()
+        expect(response.status).toEqual(200)
+    })
+
+    it('Should have three quizzes', async () => {
+        const response = await getAll()
+        expect(response.data.length).toEqual(3)
+    })
+})
